Add Windows-specific code to enable building on Windows
