+++ lib/events/sessionlog.go
@@ -17,7 +17,6 @@ limitations under the License.
 package events
 
 import (
-	"archive/tar"
 	"compress/gzip"
 	"encoding/json"
 	"fmt"
@@ -26,7 +25,6 @@ import (
 	"os"
 	"path/filepath"
 	"sync"
-	"syscall"
 	"time"
 
 	"github.com/gravitational/teleport"
@@ -136,29 +134,6 @@ func (sl *DiskSessionLogger) LogEvent(fields EventFields) error {
 // when the session logger is closed
 func (sl *DiskSessionLogger) Close() error {
 	return nil
-}
-
-func openFileForTar(filename string) (*tar.Header, io.ReadCloser, error) {
-	fi, err := os.Stat(filename)
-	if err != nil {
-		return nil, nil, trace.ConvertSystemError(err)
-	}
-	header := tar.Header{
-		Name:    filepath.Base(filename),
-		Size:    fi.Size(),
-		Mode:    int64(fi.Mode()),
-		ModTime: fi.ModTime(),
-	}
-	sys, ok := fi.Sys().(*syscall.Stat_t)
-	if ok {
-		header.Uid = int(sys.Uid)
-		header.Gid = int(sys.Gid)
-	}
-	f, err := os.Open(filename)
-	if err != nil {
-		return nil, nil, trace.ConvertSystemError(err)
-	}
-	return &header, f, nil
 }
 
 // Finalize is called by the session when it's closing. This is where we're
