+++ lib/shell/shell_native.go
@@ -1,0 +1,11 @@
+// +build !cgo
+
+package shell
+
+const (
+	DefaultShell = "/bin/sh"
+)
+
+func GetLoginShell(username string) (string, error) {
+  return DefaultShell, nil
+}
