+++ lib/utils/agentconn/agentconn_unix.go
@@ -1,0 +1,9 @@
+// +build !windows
+
+package agentconn
+
+import "net"
+
+func dialAgent(socket string) (net.Conn, error) {
+	return net.Dial("unix", socket)
+}
